closes #118, #119, #121, #122, #123
Added handling for the action "CHAMELEON_GUESS", which can be taken by
the chameleon when the game is in the state "CHAMELEON_TURN".

If the chameleon guesses the correct word, the game moves to the state
"CHAMELEON_WIN", otherwise it moves to the state "PLAYERS_WIN".

A PlayerActionResult is broadcasted to all players, with the action
result set either to "CHAMELEON_WIN" or "PLAYERS_WIN".